Reorganize mutations.
