+++ internal/syntax/parsing/delete.go
@@ -2,6 +2,7 @@ package parsing
 
 import (
 	"github.com/efritz/gostgres/internal/syntax/ast"
+	"github.com/efritz/gostgres/internal/syntax/ast/mutation"
 	"github.com/efritz/gostgres/internal/syntax/tokens"
 )
 
@@ -31,7 +32,7 @@ func (p *parser) parseDelete(token tokens.Token) (ast.BuilderResolver, error) {
 		return nil, err
 	}
 
-	return &ast.DeleteBuilder{
+	return &mutation.DeleteBuilder{
 		Target:    tableDescription,
 		Using:     usingExpressions,
 		Where:     whereExpression,
