+++ internal/syntax/parsing/insert.go
@@ -2,6 +2,7 @@ package parsing
 
 import (
 	"github.com/efritz/gostgres/internal/syntax/ast"
+	"github.com/efritz/gostgres/internal/syntax/ast/mutation"
 	"github.com/efritz/gostgres/internal/syntax/tokens"
 )
 
@@ -32,7 +33,7 @@ func (p *parser) parseInsert(token tokens.Token) (ast.BuilderResolver, error) {
 		return nil, err
 	}
 
-	return &ast.InsertBuilder{
+	return &mutation.InsertBuilder{
 		Target:      tableDescription,
 		ColumnNames: columnNames,
 		Source:      node,
